+++ changelog.md
@@ -36,6 +36,7 @@
 * [1942](https://github.com/zeta-chain/node/pull/1982) - support Bitcoin P2TR, P2WSH, P2SH, P2PKH addresses
 * [1935](https://github.com/zeta-chain/node/pull/1935) - add an operational authority group
 * [1954](https://github.com/zeta-chain/node/pull/1954) - add metric for concurrent keysigns
+* [1998](https://github.com/zeta-chain/node/pull/1998) - zetaclient multiple rpc endpoint support
 
 ### Tests